requirements.txt
Flask
requests
python-dotenv
openai==1.90.0